% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDO.R
\name{RDO}
\alias{RDO}
\title{RDO (Reproducible Data Object)}
\description{
Create and interact with Reproducible Data Objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{RDO$print()}}
\item \href{#method-new}{\code{RDO$new()}}
\item \href{#method-get_status}{\code{RDO$get_status()}}
\item \href{#method-get_name}{\code{RDO$get_name()}}
\item \href{#method-has_dependencies}{\code{RDO$has_dependencies()}}
\item \href{#method-add_dependencies}{\code{RDO$add_dependencies()}}
\item \href{#method-get_dependencies}{\code{RDO$get_dependencies()}}
\item \href{#method-get_dependency_register}{\code{RDO$get_dependency_register()}}
\item \href{#method-plot_dependencies}{\code{RDO$plot_dependencies()}}
\item \href{#method-get_code}{\code{RDO$get_code()}}
\item \href{#method-print_code}{\code{RDO$print_code()}}
\item \href{#method-run}{\code{RDO$run()}}
\item \href{#method-is_validated}{\code{RDO$is_validated()}}
\item \href{#method-invalidate}{\code{RDO$invalidate()}}
\item \href{#method-validate}{\code{RDO$validate()}}
\item \href{#method-lock}{\code{RDO$lock()}}
\item \href{#method-unlock}{\code{RDO$unlock()}}
\item \href{#method-is_locked}{\code{RDO$is_locked()}}
\item \href{#method-prune_cache}{\code{RDO$prune_cache()}}
\item \href{#method-prune_dependencies}{\code{RDO$prune_dependencies()}}
\item \href{#method-clone}{\code{RDO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Default print method of current RDO status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Other params for print function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creating a new RDO object,
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$new(name, dependencies = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Unique name of the object.}

\item{\code{dependencies}}{An RDO object or a list of RDO objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'RDO' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\subsection{Method \code{get_status()}}{
Getting current status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_name"></a>}}
\subsection{Method \code{get_name()}}{
Getting object name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$get_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character with object name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_dependencies"></a>}}
\subsection{Method \code{has_dependencies()}}{
Checking if object has dependencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$has_dependencies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if has or FALSE if not.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_dependencies"></a>}}
\subsection{Method \code{add_dependencies()}}{
Adding new or update existing RDO dependencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$add_dependencies(dependencies = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dependencies}}{An RDO object or a list of RDO objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The RDO object (self) returned invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dependencies"></a>}}
\subsection{Method \code{get_dependencies()}}{
Getting dependencies of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$get_dependencies(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function return only direct dependencies (FALSE) or
also deep indirect dependencies (dependencies of dependencies).
Default is FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of RDO dependencies wih unique names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dependency_register"></a>}}
\subsection{Method \code{get_dependency_register()}}{
Getting dependency register showing which RDO is
a direct parent for other RDO dependencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$get_dependency_register()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_dependencies"></a>}}
\subsection{Method \code{plot_dependencies()}}{
Ploting the tree of RDO dependencies.
Needs \code{visNetwork} package for plotting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$plot_dependencies()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code"></a>}}
\subsection{Method \code{get_code()}}{
Getting reproducible R code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$get_code(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function return only code for this particular RDO (FALSE)
or also from all dependencies.
Default is FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R named expression.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_code"></a>}}
\subsection{Method \code{print_code()}}{
Printing reproducible R code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$print_code(deep = FALSE, verbose = Sys.getenv("RDO_VERBOSE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function return only code for this particular RDO (FALSE)
or also from all dependencies.
Default is FALSE.}

\item{\code{verbose}}{A logical.
Should the messages be sent to console.
If the param is not set, it is read from
an environmental variable \code{RDO_VERBOSE}.
If the variable is not set, than the default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A charater. Reproducible R code returned invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
Runing reproducible R code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$run(deep = FALSE, cache = TRUE, verbose = Sys.getenv("RDO_VERBOSE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function run only code for this particular RDO (FALSE)
or should it run also dependencies' code if they are not validated.
The function is lazy and it checks
if all deep dependencies are validated first.
If so, there is no need to rerun their code again.
Default is FALSE.}

\item{\code{cache}}{A logical.
Should the result of code evaluation be cached inside an RDO object.
Default is TRUE.}

\item{\code{verbose}}{A logical.
Should the messages be sent to console.
If the param is not set, it is read from
an environmental variable \code{RDO_VERBOSE}.
If the variable is not set, than the default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The RDO object (self) returned invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_validated"></a>}}
\subsection{Method \code{is_validated()}}{
Checking if RDO is validated.
A RDO is validated when the result of running reproducible R code
saved inside the RDO is the same as data cached inside the RDO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$is_validated(deep = FALSE, verbose = Sys.getenv("RDO_VERBOSE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function validate only this particular RDO (FALSE)
or should it validate also all deep dependencies (TRUE).
Default is FALSE.}

\item{\code{verbose}}{A logical.
Should the messages be sent to console.
If the param is not set, it is read from
an environmental variable \code{RDO_VERBOSE}.
If the variable is not set, than the default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical. TRUE if RDO is validated, FALSE if not.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate"></a>}}
\subsection{Method \code{invalidate()}}{
Invalidating the RDO explicitly
by setting the 'is_validated' status to 'FALSE'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$invalidate(deep = FALSE, verbose = Sys.getenv("RDO_VERBOSE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function invalidate only this particular RDO (FALSE)
or should it invalidate also all deep dependencies (TRUE).
Default is FALSE.}

\item{\code{verbose}}{A logical.
Should the messages be sent to console.
If the param is not set, it is read from
an environmental variable \code{RDO_VERBOSE}.
If the variable is not set, than the default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The RDO object (self) returned invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\subsection{Method \code{validate()}}{
Validating the RDO explicitly
by running deep reproducible R code
and checking if the result is the same as cached data inside the RDO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$validate(deep = FALSE, verbose = Sys.getenv("RDO_VERBOSE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function validate only this particular RDO (FALSE)
or should it validate also all deep dependencies (TRUE).
Default is FALSE.}

\item{\code{verbose}}{A logical.
Should the messages be sent to console.
If the param is not set, it is read from
an environmental variable \code{RDO_VERBOSE}.
If the variable is not set, than the default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The RDO object (self) returned invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lock"></a>}}
\subsection{Method \code{lock()}}{
Locking the RDO object.
When RDO object is locked you
cannot change the R code and data cache saved inside the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$lock(deep = FALSE, verbose = Sys.getenv("RDO_VERBOSE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function lock only this particular RDO (FALSE)
or should it lock also all deep dependencies (TRUE).
Default is FALSE.}

\item{\code{verbose}}{A logical.
Should the messages be sent to console.
If the param is not set, it is read from
an environmental variable \code{RDO_VERBOSE}.
If the variable is not set, than the default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The RDO object (self) returned invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unlock"></a>}}
\subsection{Method \code{unlock()}}{
Unlocking previously locked RDO object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$unlock(deep = FALSE, verbose = Sys.getenv("RDO_VERBOSE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function unlock only this particular RDO (FALSE)
or should it unlock also all deep dependencies (TRUE).
Default is FALSE.}

\item{\code{verbose}}{A logical.
Should the messages be sent to console.
If the param is not set, it is read from
an environmental variable \code{RDO_VERBOSE}.
If the variable is not set, than the default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The RDO object (self) returned invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_locked"></a>}}
\subsection{Method \code{is_locked()}}{
Checking if an RDO object is locked.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$is_locked(deep = FALSE, verbose = Sys.getenv("RDO_VERBOSE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function check only this particular RDO (FALSE)
or should it check also all deep dependencies (TRUE).
Default is FALSE.}

\item{\code{verbose}}{A logical.
Should the messages be sent to console.
If the param is not set, it is read from
an environmental variable \code{RDO_VERBOSE}.
If the variable is not set, than the default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical. TRUE if RDO is locked, FALSE if not.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prune_cache"></a>}}
\subsection{Method \code{prune_cache()}}{
Prunning (clearing) RDO data cache by setting
cache to NULL.
It can save memory when we no longer need
to keep cache in depencencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$prune_cache(deep = FALSE, verbose = Sys.getenv("RDO_VERBOSE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{A logical.
Should the function prune cache of only this particular RDO (FALSE)
or should it prune cache also of deep dependencies (TRUE).
If an RDO is locked the cache in this particular RDO is not pruned.
Default is FALSE.}

\item{\code{verbose}}{A logical.
Should the messages be sent to console.
If the param is not set, it is read from
an environmental variable \code{RDO_VERBOSE}.
If the variable is not set, than the default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The RDO object (self) returned invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prune_dependencies"></a>}}
\subsection{Method \code{prune_dependencies()}}{
Prunning RDO dependencies by ensuring that RDO objects
in deep dependencies with the same name point to the same RDO objects.
This type of prunning may be useful after deep clonning of
complex RDO tree with duplicated dependencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$prune_dependencies(verbose = Sys.getenv("RDO_VERBOSE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{A logical.
Should the messages be sent to console.
If the param is not set, it is read from
an environmental variable \code{RDO_VERBOSE}.
If the variable is not set, than the default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The RDO object (self) returned invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
